requirements.txt
# Vercel Python Runtime Dependencies
# ────────────────────────────────────
# This file is used by Vercel to install Python dependencies for the
# serverless API function. Keep it minimal for fast cold starts.

fastapi==0.110.1
uvicorn==0.25.0
mangum>=0.17.0
pydantic==2.12.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
PyJWT==2.11.0
supabase>=2.0.0
aiohttp==3.13.3
numpy==2.4.2
pandas==3.0.1
yfinance>=0.2.40
anthropic>=0.49.0
starlette==0.37.2
